requirements.txt
aiogram==3.4.1
pydantic==2.5.3
pydantic-core==2.14.6
fastapi==0.110.0
uvicorn==0.27.1
gunicorn==21.2.0
python-dotenv==1.0.1
